fix: move /workshops to /quest
