+++ app/quest/components/WorkshopSlotElement.tsx
