+++ next.config.ts
@@ -4,6 +4,9 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"],
   output: "standalone",
+  redirects: async () => [
+    { source: "/workshops", destination: "/quest", permanent: true },
+  ],
 };
 
 export default createMDX()(nextConfig);
